+++ COMETwebapp/ViewModels/Components/ModelEditor/ElementDefinitionCreationViewModel.cs
@@ -1,15 +1,26 @@
-﻿// --------------------------------------------------------------------------------------------------------
-// <copyright file="ParticipantCreationViewModel.cs" company="RHEA System S.A.">
-//  Copyright (c) 2022 RHEA System S.A.
+﻿// --------------------------------------------------------------------------------------------------------------------
+//  <copyright file="ElementDefinitionCreationViewModel.cs" company="RHEA System S.A.">
+//     Copyright (c) 2023 RHEA System S.A.
 // 
-//  Author: Antoine Théate, Sam Gerené, Alex Vorobiev, Alexander van Delft, Martin Risseeuw
+//     Authors: Sam Gerené, Alex Vorobiev, Alexander van Delft, Jaime Bernar, Théate Antoine
 // 
-//  This file is part of UI-DSM.
-//  The UI-DSM web application is used to review an ECSS-E-TM-10-25 model.
+//     This file is part of COMET WEB Community Edition
+//     The COMET WEB Community Edition is the RHEA Web Application implementation of ECSS-E-TM-10-25 Annex A and Annex C.
 // 
-//  The UI-DSM application is provided to the community under the Apache License 2.0.
-// </copyright>
-// --------------------------------------------------------------------------------------------------------
+//     The COMET WEB Community Edition is free software; you can redistribute it and/or
+//     modify it under the terms of the GNU Affero General Public
+//     License as published by the Free Software Foundation; either
+//     version 3 of the License, or (at your option) any later version.
+// 
+//     The COMET WEB Community Edition is distributed in the hope that it will be useful,
+//     but WITHOUT ANY WARRANTY; without even the implied warranty of
+//     MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the GNU
+//    Affero General Public License for more details.
+// 
+//    You should have received a copy of the GNU Affero General Public License
+//    along with this program.  If not, see <http://www.gnu.org/licenses/>.
+//  </copyright>
+//  --------------------------------------------------------------------------------------------------------------------
 
 namespace COMETwebapp.ViewModels.Components.ModelEditor
 {
